More changes requested
Changes needed were to make sure that on the small 4 inch device screens that the UI wouldn't get smashed.

Other changes as requested is to ensure that we are using a two case enum for our state with a seperate function that takes in a parameter for the state to feed a switch statement in to re-use some functionality.